# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A backend service written in Go with Gin framework",
        "title": "LONDONWINNER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/transactions": {
            "get": {
                "description": "Retrieve all transactions with optional pagination and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get all transactions with pagination and filtering",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "pageNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 format)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 format)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Telecom provider",
                        "name": "telecom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Agent code",
                        "name": "agentCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by phone number",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginationResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/activators/stats": {
            "get": {
                "description": "Retrieves user transaction stats grouped by day",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get daily user transaction stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/all-tokens/{phoneNumber}": {
            "get": {
                "description": "Retrieves ongoing, won, and expired tokens for a given phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get all tokens for a phone number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phoneNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/getTokenByPhoneNumber": {
            "get": {
                "description": "Retrieves all transactions and their associated tokens for a given phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transactions and tokens by phone number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number of the customer",
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TransactionTokenResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Phone number is missing",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/initialize": {
            "post": {
                "description": "Initializes payment transaction request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Initialize a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "apiKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Transaction data",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTransactionDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/transactions/my-token-stats/{phoneNumber}": {
            "get": {
                "description": "Retrieves token statistics grouped by product for a given phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get token statistics for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number of the user",
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/my-tokens/{phoneNumber}": {
            "get": {
                "description": "Retrieves ongoing, won, and expired tokens for a given phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get all tokens for a phone number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phoneNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/payment/request": {
            "post": {
                "description": "Processes a payment request and sends Kafka event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Process a payment request",
                "parameters": [
                    {
                        "description": "Payment data",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PaymentDto"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/phone_number-hits": {
            "get": {
                "description": "Retrieves stats on transactions grouped by phone number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transaction stats by phone number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "pageNumber",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/player-stats": {
            "get": {
                "description": "Retrieve player statistics with optional date range filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get player statistics",
                "parameters": [
                    {
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "type": "string",
                        "description": "Date range (daily, weekly, monthly)",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PlayerStats"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/product-stats": {
            "get": {
                "description": "Retrieve product statistics with optional date range filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get product statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 format)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 format)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "type": "string",
                        "description": "Date range (daily, weekly, monthly)",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProductStats"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/regenerate/token/{referenceId}": {
            "post": {
                "description": "Sends a request to an external API to regenerate a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Regenerate Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "referenceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/resend-token/{token}": {
            "post": {
                "description": "Resends a token via SMS for the given token ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Resend Token SMS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/transactions/{referenceId}/status": {
            "get": {
                "description": "Returns the status and details of a transaction by its reference ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Check the status of a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "apiKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reference ID of the transaction",
                        "name": "referenceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.CreateTransactionDto": {
            "type": "object",
            "required": [
                "amount",
                "phone_number",
                "productId",
                "reference_id",
                "telco",
                "transaction_date"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 0
                },
                "phone_number": {
                    "type": "string",
                    "example": "250788205965"
                },
                "productId": {
                    "type": "string",
                    "example": "string"
                },
                "reference_id": {
                    "type": "string",
                    "example": "string"
                },
                "telco": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Telco"
                        }
                    ],
                    "example": "string"
                },
                "token": {
                    "type": "string",
                    "example": "string"
                },
                "transaction_date": {
                    "type": "string",
                    "example": "2025-02-18T08:45:19.803Z"
                }
            }
        },
...
        "models.TransactionTokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/models.Token"
                },
                "transaction": {
                    "$ref": "#/definitions/models.Transaction"
                }
            }
        },
        "utils.PaginationResponse": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "data": {},
                "lastPage": {
                    "type": "integer"
                },
                "nextPage": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "previousPage": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}